aurax-ocr-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
paddleocr==2.7.3
paddlepaddle==2.6.2
httpx==0.26.0
pillow==10.2.0
numpy==1.24.3
pydantic==2.5.3
python-multipart==0.0.6
